server/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
coloredlogs==10.0
Flask==1.1.1
Flask-Cors==3.0.8
google==2.0.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-automl==0.5.0
google-cloud-core==1.0.3
google-cloud-datastore==1.9.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
humanfriendly==4.18
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
micropython-cpython-ubinascii==0.0
nltk==3.4.5
numpy==1.17.2
pillow==6.1.0
protobuf==3.9.2
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pyreadline==2.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.8
urllib3==1.25.5
Werkzeug==0.15.5
yaml-1.3==0.1.0